refactor: excel-download (DEW+v).
