+++ src/OCADataValidator/useHandleJsonDrop.js
@@ -3,10 +3,19 @@ import { messages } from "../constants/messages";
 import { Context } from "../App";
 import useZipParser from "../StartSchema/useZipParser";
 import JSZip from "jszip";
+import { json } from "react-router-dom";
 
-const neededOverlays = ['format', 'character_encoding', 'conformance', 'entry_code'];
+const neededOverlays = [
+  "format",
+  "character_encoding",
+  "conformance",
+  "entry_code",
+];
 
-export const useHandleJsonDrop = (setShowWarningCard, firstTimeDisplayWarning) => {
+export const useHandleJsonDrop = (
+  setShowWarningCard,
+  firstTimeDisplayWarning
+) => {
   const {
     setCurrentDataValidatorPage,
     setZipToReadme,
@@ -23,15 +32,20 @@ export const useHandleJsonDrop = (setShowWarningCard, firstTimeDisplayWarning) =
     datasetRawFile,
     setMatchingRowData,
     setJsonParsedFile,
-    firstTimeMatchingRef
+    firstTimeMatchingRef,
+    targetResult,
+    setTargetResult,
   } = useContext(Context);
   const {
     processLanguages,
     processMetadata,
-    processLabelsDescriptionRootUnitsEntries
+    processLabelsDescriptionRootUnitsEntries,
   } = useZipParser();
 
-  const [jsonDropMessage, setJsonDropMessage] = useState({ message: "", type: "" });
+  const [jsonDropMessage, setJsonDropMessage] = useState({
+    message: "",
+    type: "",
+  });
 
   const overallLoading = useCallback(() => {
     setJsonLoading(true);
@@ -49,180 +63,205 @@ export const useHandleJsonDrop = (setShowWarningCard, firstTimeDisplayWarning) =
     setShowWarningCard(false);
   }, []);
 
-  const handleJsonDrop = useCallback((acceptedFiles) => {
-    try {
-      setJsonLoading(true);
-      const reader = new FileReader();
+  const handleJsonDrop = useCallback(
+    (acceptedFiles) => {
+      try {
+        setJsonLoading(true);
+        const reader = new FileReader();
 
-      reader.onload = async (e) => {
-        const rawParse = JSON.parse(e.target.result);
-        let jsonFile = null;
-        if (rawParse?.['bundle']) {
-          jsonFile = rawParse?.['bundle'];
-        } else if (rawParse?.['schema']?.[0]) {
-          jsonFile = rawParse?.['schema']?.[0];
-        } else {
-          jsonFile = rawParse;
-        }
-
-        if (!jsonFile) {
-          throw new Error('No JSON file found');
-        }
-
-        setJsonParsedFile(jsonFile);
-        const languageList = [];
-        const informationList = [];
-        const labelList = [];
-        const metaList = [];
-        const entryList = [];
-        const allJSONFiles = [];
-        let loadRoot = undefined;
-        let entryCodeSummary = {};
-        let conformance = undefined;
-        let characterEncoding = undefined;
-        let loadUnits = undefined;
-        let formatRules = undefined;
-        let cardinalityData = undefined;
-
-        // load up metadata file in OCA bundle
-        if (jsonFile?.overlays?.meta) {
-          metaList.push(...jsonFile.overlays.meta);
-          languageList.push(...jsonFile.overlays.meta.map((meta) => meta.language.slice(0, 2)));
-
-          // ONLY for README
-          const readmeMeta = jsonFile.overlays.meta.map((meta) => {
-            return JSON.stringify(meta);
-          });
-          allJSONFiles.push(...readmeMeta);
-        }
-
-        if (jsonFile?.overlays?.information) {
-          informationList.push(...jsonFile.overlays.information);
-
-          // ONLY for README
-          const readmeInformation = jsonFile.overlays.information.map((information) => {
-            return JSON.stringify(information);
-          });
-          allJSONFiles.push(...readmeInformation);
-        }
-
-        if (jsonFile?.overlays?.label) {
-          labelList.push(...jsonFile.overlays.label);
-
-          // ONLY for README
-          const readmeLabel = jsonFile.overlays.label.map((label) => {
-            return JSON.stringify(label);
-          });
-          allJSONFiles.push(...readmeLabel);
-        }
-
-        if (jsonFile?.['capture_base']) {
-          if (jsonFile?.['capture_base']?.['flagged_attributes']?.length > 0) {
-            setShowWarningCard(true);
+        reader.onload = async (e) => {
+          setTargetResult(e);
+          const textDecoder = new TextDecoder("utf-8");
+          const jsonString = textDecoder.decode(e.target.result);
+          const rawParse = JSON.parse(jsonString);
+          let jsonFile = null;
+          if (rawParse?.["bundle"]) {
+            jsonFile = rawParse?.["bundle"];
+          } else if (rawParse?.["schema"]?.[0]) {
+            jsonFile = rawParse?.["schema"]?.[0];
+          } else {
+            jsonFile = rawParse;
           }
-          loadRoot = { ...jsonFile['capture_base'] };
 
-          // ONLY for README
-          allJSONFiles.push(JSON.stringify(loadRoot));
+          if (!jsonFile) {
+            throw new Error("No JSON file found");
+          }
+
+          setJsonParsedFile(jsonFile);
+          const languageList = [];
+          const informationList = [];
+          const labelList = [];
+          const metaList = [];
+          const entryList = [];
+          const allJSONFiles = [];
+          let loadRoot = undefined;
+          let entryCodeSummary = {};
+          let conformance = undefined;
+          let characterEncoding = undefined;
+          let loadUnits = undefined;
+          let formatRules = undefined;
+          let cardinalityData = undefined;
+
+          // load up metadata file in OCA bundle
+          if (jsonFile?.overlays?.meta) {
+            metaList.push(...jsonFile.overlays.meta);
+            languageList.push(
+              ...jsonFile.overlays.meta.map((meta) => meta.language.slice(0, 2))
+            );
+
+            // ONLY for README
+            const readmeMeta = jsonFile.overlays.meta.map((meta) => {
+              return JSON.stringify(meta);
+            });
+            allJSONFiles.push(...readmeMeta);
+          }
+
+          if (jsonFile?.overlays?.information) {
+            informationList.push(...jsonFile.overlays.information);
+
+            // ONLY for README
+            const readmeInformation = jsonFile.overlays.information.map(
+              (information) => {
+                return JSON.stringify(information);
+              }
+            );
+            allJSONFiles.push(...readmeInformation);
+          }
+
+          if (jsonFile?.overlays?.label) {
+            labelList.push(...jsonFile.overlays.label);
+
+            // ONLY for README
+            const readmeLabel = jsonFile.overlays.label.map((label) => {
+              return JSON.stringify(label);
+            });
+            allJSONFiles.push(...readmeLabel);
+          }
+
+          if (jsonFile?.["capture_base"]) {
+            if (
+              jsonFile?.["capture_base"]?.["flagged_attributes"]?.length > 0
+            ) {
+              setShowWarningCard(true);
+            }
+            loadRoot = { ...jsonFile["capture_base"] };
+
+            // ONLY for README
+            allJSONFiles.push(JSON.stringify(loadRoot));
+          }
+
+          if (jsonFile?.overlays?.unit) {
+            loadUnits = { ...jsonFile.overlays.unit };
+
+            // ONLY for README
+            allJSONFiles.push(JSON.stringify(loadUnits));
+          }
+
+          if (jsonFile?.overlays?.conformance) {
+            conformance = { ...jsonFile.overlays.conformance };
+
+            // ONLY for README
+            allJSONFiles.push(JSON.stringify(conformance));
+          }
+
+          if (jsonFile?.overlays?.["character_encoding"]) {
+            characterEncoding = { ...jsonFile.overlays["character_encoding"] };
+
+            // ONLY for README
+            allJSONFiles.push(JSON.stringify(characterEncoding));
+          }
+
+          if (jsonFile?.overlays?.entry_code) {
+            entryCodeSummary = { ...jsonFile.overlays.entry_code };
+
+            // ONLY for README
+            allJSONFiles.push(JSON.stringify(entryCodeSummary));
+          }
+
+          if (jsonFile?.overlays?.["format"]) {
+            formatRules = { ...jsonFile.overlays["format"] };
+
+            // ONLY for README
+            allJSONFiles.push(JSON.stringify(formatRules));
+          }
+
+          if (jsonFile?.overlays?.entry) {
+            entryList.push(...jsonFile.overlays.entry);
+
+            // ONLY for README
+            const readmeEntry = jsonFile.overlays.entry.map((entry) => {
+              return JSON.stringify(entry);
+            });
+            allJSONFiles.push(...readmeEntry);
+          }
+
+          if (jsonFile?.overlays?.["cardinality"]) {
+            cardinalityData = { ...jsonFile.overlays["cardinality"] };
+
+            // ONLY for README
+            allJSONFiles.push(JSON.stringify(cardinalityData));
+          }
+
+          if (!languageList || languageList.length === 0) {
+            throw new Error("No language found in the JSON file");
+          }
+
+          processLanguages(languageList);
+          processMetadata(metaList);
+          processLabelsDescriptionRootUnitsEntries(
+            labelList,
+            informationList,
+            loadRoot,
+            loadUnits,
+            entryCodeSummary,
+            entryList,
+            conformance,
+            characterEncoding,
+            languageList,
+            formatRules,
+            cardinalityData
+          );
+          setZipToReadme(allJSONFiles);
+        };
+
+        reader.readAsArrayBuffer(acceptedFiles[0]);
+
+        reader.onloadend = () => {
+          setTimeout(() => {
+            setJsonDropDisabled(true);
+            setJsonDropMessage({ message: "", type: "" });
+            setJsonLoading(false);
+            setDatasetLoading(false);
+            if (datasetRawFile.length === 0) {
+              setDatasetDropDisabled(false);
+            }
+            if (!jsonIsParsed) {
+              setJsonIsParsed(true);
+              setCurrentDataValidatorPage("SchemaViewDataValidator");
+            }
+          }, 900);
+        };
+      } catch (error) {
+        setJsonDropMessage({ message: messages.uploadFail, type: "error" });
+        setJsonLoading(false);
+        setDatasetLoading(false);
+        if (datasetRawFile.length === 0) {
+          setDatasetDropDisabled(false);
         }
-
-        if (jsonFile?.overlays?.unit) {
-          loadUnits = { ...jsonFile.overlays.unit };
-
-          // ONLY for README
-          allJSONFiles.push(JSON.stringify(loadUnits));
-        }
-
-        if (jsonFile?.overlays?.conformance) {
-          conformance = { ...jsonFile.overlays.conformance };
-
-          // ONLY for README
-          allJSONFiles.push(JSON.stringify(conformance));
-        }
-
-        if (jsonFile?.overlays?.['character_encoding']) {
-          characterEncoding = { ...jsonFile.overlays['character_encoding'] };
-
-          // ONLY for README
-          allJSONFiles.push(JSON.stringify(characterEncoding));
-        }
-
-        if (jsonFile?.overlays?.entry_code) {
-          entryCodeSummary = { ...jsonFile.overlays.entry_code };
-
-          // ONLY for README
-          allJSONFiles.push(JSON.stringify(entryCodeSummary));
-        }
-
-        if (jsonFile?.overlays?.['format']) {
-          formatRules = { ...jsonFile.overlays['format'] };
-
-          // ONLY for README
-          allJSONFiles.push(JSON.stringify(formatRules));
-        }
-
-        if (jsonFile?.overlays?.entry) {
-          entryList.push(...jsonFile.overlays.entry);
-
-          // ONLY for README
-          const readmeEntry = jsonFile.overlays.entry.map((entry) => {
-            return JSON.stringify(entry);
-          });
-          allJSONFiles.push(...readmeEntry);
-        }
-
-        if (jsonFile?.overlays?.['cardinality']) {
-          cardinalityData = { ...jsonFile.overlays['cardinality'] };
-
-          // ONLY for README
-          allJSONFiles.push(JSON.stringify(cardinalityData));
-        }
-
-        if (!languageList || languageList.length === 0) {
-          throw new Error('No language found in the JSON file');
-        }
-
-        processLanguages(languageList);
-        processMetadata(metaList);
-        processLabelsDescriptionRootUnitsEntries(labelList, informationList, loadRoot, loadUnits, entryCodeSummary, entryList, conformance, characterEncoding, languageList, formatRules, cardinalityData);
-        setZipToReadme(allJSONFiles);
-      };
-
-      reader.readAsText(acceptedFiles[0]);
-
-      reader.onloadend = () => {
         setTimeout(() => {
-          setJsonDropDisabled(true);
           setJsonDropMessage({ message: "", type: "" });
-          setJsonLoading(false);
-          setDatasetLoading(false);
-          if (datasetRawFile.length === 0) {
-            setDatasetDropDisabled(false);
-          }
-          if (!jsonIsParsed) {
-            setJsonIsParsed(true);
-            setCurrentDataValidatorPage("SchemaViewDataValidator");
-          }
-        }, 900);
-      };
-    } catch (error) {
-      setJsonDropMessage({ message: messages.uploadFail, type: "error" });
-      setJsonLoading(false);
-      setDatasetLoading(false);
-      if (datasetRawFile.length === 0) {
-        setDatasetDropDisabled(false);
+        }, [2500]);
       }
-      setTimeout(() => {
-        setJsonDropMessage({ message: "", type: "" });
-      }, [2500]);
-    }
-  }, [datasetRawFile.length, jsonIsParsed]);
+    },
+    [datasetRawFile.length, jsonIsParsed]
+  );
 
   const handleZipDrop = useCallback((acceptedFiles) => {
     try {
       const reader = new FileReader();
 
       reader.onload = async (e) => {
+        setTargetResult(e);
         const zip = await JSZip.loadAsync(e.target.result);
         const languageList = [];
         const informationList = [];
@@ -246,11 +285,15 @@ export const useHandleJsonDrop = (setShowWarningCard, firstTimeDisplayWarning) =
 
         // loop through all files in OCA bundle
         for (const [key, file] of Object.entries(metadataJson.files[root])) {
-          const content = await zip.files[file + '.json'].async("text");
+          const content = await zip.files[file + ".json"].async("text");
           const parsedContent = JSON.parse(content);
 
-          if ('type' in parsedContent && neededOverlays.includes(parsedContent['type'].split('/')[2])) {
-            bundleForValidator.overlays[parsedContent['type'].split('/')[2]] = parsedContent;
+          if (
+            "type" in parsedContent &&
+            neededOverlays.includes(parsedContent["type"].split("/")[2])
+          ) {
+            bundleForValidator.overlays[parsedContent["type"].split("/")[2]] =
+              parsedContent;
           }
 
           if (key.includes("meta")) {
@@ -264,7 +307,6 @@ export const useHandleJsonDrop = (setShowWarningCard, firstTimeDisplayWarning) =
             // Format word is inside Information word, so we need to check if it is a format or information
             formatRules = parsedContent;
           }
-
 
           if (key.includes("label")) {
             labelList.push(parsedContent);
@@ -297,20 +339,37 @@ export const useHandleJsonDrop = (setShowWarningCard, firstTimeDisplayWarning) =
           allZipFiles.push(content);
         }
 
-        const loadRoot = await zip.files[metadataJson.root + '.json'].async("text");
+        const loadRoot = await zip.files[metadataJson.root + ".json"].async(
+          "text"
+        );
         const parsedRoot = JSON.parse(loadRoot);
-        if (parsedRoot?.['flagged_attributes']?.length > 0) {
+        if (parsedRoot?.["flagged_attributes"]?.length > 0) {
           setShowWarningCard(true);
         }
-        if ('type' in parsedRoot && parsedRoot['type'].split('/')[1] === 'capture_base') {
-          bundleForValidator['capture_base'] = parsedRoot;
+        if (
+          "type" in parsedRoot &&
+          parsedRoot["type"].split("/")[1] === "capture_base"
+        ) {
+          bundleForValidator["capture_base"] = parsedRoot;
         }
         allZipFiles.push(loadRoot);
 
         setJsonParsedFile(bundleForValidator);
         processLanguages(languageList);
         processMetadata(metaList);
-        processLabelsDescriptionRootUnitsEntries(labelList, informationList, JSON.parse(loadRoot), loadUnits, entryCodeSummary, entryList, conformance, characterEncoding, languageList, formatRules, cardinalityData);
+        processLabelsDescriptionRootUnitsEntries(
+          labelList,
+          informationList,
+          JSON.parse(loadRoot),
+          loadUnits,
+          entryCodeSummary,
+          entryList,
+          conformance,
+          characterEncoding,
+          languageList,
+          formatRules,
+          cardinalityData
+        );
         setZipToReadme(allZipFiles);
       };
 
@@ -345,9 +404,17 @@ export const useHandleJsonDrop = (setShowWarningCard, firstTimeDisplayWarning) =
   }, []);
 
   useEffect(() => {
-    if (jsonRawFile && jsonRawFile.length > 0 && jsonRawFile[0].path.includes(".json")) {
+    if (
+      jsonRawFile &&
+      jsonRawFile.length > 0 &&
+      jsonRawFile[0].path.includes(".json")
+    ) {
       handleJsonDrop(jsonRawFile);
-    } else if (jsonRawFile && jsonRawFile.length > 0 && jsonRawFile[0].path.includes(".zip")) {
+    } else if (
+      jsonRawFile &&
+      jsonRawFile.length > 0 &&
+      jsonRawFile[0].path.includes(".zip")
+    ) {
       handleZipDrop(jsonRawFile);
     } else if (jsonRawFile && jsonRawFile.length > 0) {
       setJsonDropMessage({ message: messages.uploadFail, type: "error" });
@@ -371,6 +438,7 @@ export const useHandleJsonDrop = (setShowWarningCard, firstTimeDisplayWarning) =
     jsonDropMessage,
     setJsonDropMessage,
     setCurrentDataValidatorPage,
-    handleClearJSON
+    handleClearJSON,
+    targetResult,
   };
 };
